+++ src/main.js
@@ -2,14 +2,10 @@ import TripInfoComponent from '@/components/trip-info';
 import PriceInfoComponent from '@/components/price-info';
 import TripTabsComponent from '@/components/trip-tabs';
 import FiltersComponent from '@/components/trip-filters';
-import SortComponent from '@/components/trip-sort';
-import EditEventComponent from '@/components/trip-edit';
-import DaysListComponent from '@/components/days-list';
-import DayComponent from '@/components/trip-day';
-import EventComponent from '@/components/trip-event';
+
 import {generateEvents} from '@/mock/events';
-import {splitEventsByDays} from '@/components/sort';
-import {RenderPosition, replace, render} from '@/utils/render';
+import {RenderPosition, render} from '@/utils/render';
+import TripController from '@/controllers/trip-controller';
 
 const EVENTS_AMOUNT = 20;
 
@@ -33,52 +29,5 @@ render(tripControlsElement, new FiltersComponent(), RenderPosition.BEFOREEND);
 
 const tripEventsElement = document.querySelector(`.trip-events`);
 
-// sorting line
-render(tripEventsElement, new SortComponent(), RenderPosition.BEFOREEND);
-// days and events container
-render(tripEventsElement, new DaysListComponent(), RenderPosition.BEFOREEND);
-
-// days and events
-const [eventsByDays, dates] = splitEventsByDays(events);
-
-// render days column
-const tripDaysListElement = tripEventsElement.querySelector(`.trip-days`);
-
-dates.forEach((day, i, arr) => {
-  const getDifferenceInDays = (start, end) => {
-    const MS_IN_DAY = 86400000;
-
-    return Math.floor((end - start) / MS_IN_DAY);
-  };
-
-  const counter = getDifferenceInDays(arr[0], day);
-
-  render(tripDaysListElement, new DayComponent(day, counter + 1), RenderPosition.BEFOREEND);
-});
-
-const renderEvent = (eventListElement, event) => {
-  const rollupButtonClickHandler = () => {
-    replace(editEventComponent, eventComponent);
-  };
-
-  const editFormSubmitHandler = () => {
-    replace(eventComponent, editEventComponent);
-  };
-
-  const eventComponent = new EventComponent(event);
-  const editEventComponent = new EditEventComponent(event);
-
-  eventComponent.setRollupButtonClickHandler(rollupButtonClickHandler);
-  editEventComponent.setSubmitHandler(editFormSubmitHandler);
-
-  render(eventListElement, eventComponent, RenderPosition.BEFOREEND);
-};
-
-// render events inside of days
-const tripEventsListElements = tripDaysListElement.querySelectorAll(`.trip-events__list`);
-
-tripEventsListElements.forEach((it, i) => {
-  eventsByDays[i].forEach((event) => {
-    renderEvent(it, event);
-  });
-});
+const tripController = new TripController(tripEventsElement);
+tripController.render(events);
